+++ CHANGELOG.md
@@ -81,3 +81,11 @@
 - Creación y cambio a la rama `Sprint_2`.
 - Creación de la estructura de directorios para la base de datos relacional (`database`, `models`, `migrations/sql`).
 - Creación de archivos base vacíos (`connection.py`, `models.py`, `migrations.py`).
+
+## [2026-05-21]
+
+### Añadido
+#### Conexión a Base de Datos (Ejercicio 02)
+- Creación del archivo `price_manager/database/connection.py`.
+- Implementación de la clase `ConexionDB` utilizando `SQLAlchemy` con motor SQLite.
+- Configuración para la creación automática del directorio de la base de datos local (`/content/base_de_datos/`).